* Add map_watch_interval configurable. Support floating point values for time intervals.

+++ src/cfg_file.h
@@ -409,7 +409,7 @@ guint64 parse_limit (const gchar *limit, guint len);
  * @param default_type dimension of time if no suffix is specified
  * @return value of time in milliseconds
  */
-guint cfg_parse_time (const gchar *t, enum time_type default_type);
+gdouble cfg_parse_time (const gchar *t, enum time_type default_type);
 
 /**
  * Parse flag
